+++ Hestia/Hestia/Backenda/Exceptions/ServerInteractionException.cs
