+++ Hestia/Hestia/Backenda/Models/Deserializers/DeviceDeserializer.cs
