+++ Hestia/Hestia/Backenda/Models/Device.cs
