+++ Hestia/Hestia/Backenda/Models/PluginInfo.cs
