+++ Hestia/Hestia/Backenda/Utils/ServerDiscovery/ServerDelegate.cs
