--- Hestia/Hestia/backend/authentication/Auth0Connector.cs
@@ -1,19 +1,0 @@
-﻿using Auth0.OidcClient;
-using UIKit;
-
-namespace Hestia.Backend.Authentication
-{
-    public static class Auth0Connector
-    {
-        public static Auth0Client CreateAuthClient(UIViewController uIViewController)
-        {
-            Auth0Client client = new Auth0Client(new Auth0ClientOptions
-            {
-                Domain = Resources.strings.domainAuth0,
-                ClientId = Resources.strings.clientId,
-                Controller = uIViewController
-            });
-            return client;
-        }
-    }
-}
